Move scoring envs to dedicated module
Refactor: extract ScoringEnvs and prepare_scoring_envs() from config_utils.py into a new scoring_envs.py and move prepare_crispron_env() into config_crispron.py. Update imports in __main__.py and config_crispron.py to use the new module, add platform checks/warnings and create_mamba_env usage for CRISPRon env setup, and tidy up imports (remove unused zipfile/shutil from utils.py). This separates scoring environment setup from generic config utilities and simplifies responsibilities.
